aggiunti esercizi

+++ code/build.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 set -e
-cp .env.example .env
+if test -f "games.bson"; then
+    echo ".env already exists"
+else
+    cp .env.example .env
+fi
 docker-compose build
 npm install
